fix staking tests
