+++ platform/cosmos/model.go
@@ -100,6 +100,10 @@ type Inflation struct {
 	Result string `json:"result"`
 }
 
+type Delegations struct {
+	List []Delegation `json:"result"`
+}
+
 type Delegation struct {
 	DelegatorAddress string `json:"delegator_address"`
 	ValidatorAddress string `json:"validator_address"`
@@ -114,6 +118,10 @@ func (d *Delegation) Value() string {
 	return d.Shares
 }
 
+type UnbondingDelegations struct {
+	List []UnbondingDelegation `json:"result"`
+}
+
 type UnbondingDelegation struct {
 	Delegation
 	Entries []UnbondingDelegationEntry `json:"entries"`
@@ -126,6 +134,10 @@ type UnbondingDelegationEntry struct {
 }
 
 type StakingPool struct {
+	Pool Pool `json:"result"`
+}
+
+type Pool struct {
 	NotBondedTokens string `json:"not_bonded_tokens"`
 	BondedTokens    string `json:"bonded_tokens"`
 }
